Move create post form to a separate file
